+++ src/components/bootstrap/dropdown/dropdown.spec.ts
@@ -1,5 +1,5 @@
 import {Component, DebugElement} from '@angular/core';
-import {TestBed, async, ComponentFixture} from '@angular/core/testing';
+import {TestBed, ComponentFixture} from '@angular/core/testing';
 import {By} from '@angular/platform-browser';
 
 import {isDefined} from 'caesium-core/lang';
@@ -16,14 +16,13 @@ import {CsDropdown, CsDropdownToggle} from './dropdown';
     </style>
 
     <cs-dropdown id="hosted-dropdown" (open$)="handleOpen($event)">
-        <button csDropdownToggle id="dropdown-toggle">Dropdown</button>
+        <button csDropdownToggle class="btn" id="dropdown-toggle">Dropdown</button>
         
         <div id="dropdown-content" class="content" >
             Lorem ipsum dolor amet
         </div>
     </cs-dropdown>
     
-    <cs-dropdown id="open-dropdown" (open$)="handleOpen($event)"></cs-dropdown>
     `,
 })
 export class DropdownHost {
@@ -34,39 +33,39 @@ export class DropdownHost {
 export function expectIsDropdownOpen(dropdown: DebugElement, open?: boolean) {
     open = !isDefined(open) || open;
 
-    console.log('shadow root', dropdown.nativeElement.shadowRoot);
-    let content = dropdown.nativeElement.shadowRoot.querySelector('#dropdown-content');
+    let content: DebugElement = dropdown.query(By.css('#dropdown-content'));
+    expect(content).not.toBeNull('#dropdown-content should be distributed into the node');
+    expect(content.nativeElement.textContent)
+            .toMatch('Lorem ipsum dolor amet', '#dropdown-content should have existing content');
     if (open) {
-        expect(content).not.toBeNull('#dropdown-content should be distributed into the node');
-        expect(content.textContent).toMatch('Lorem ipsum dolor amet', '#dropdown-content should have existing content');
+        expect(dropdown.nativeElement.classList).toContain('open');
     } else {
-        expect(content).toBeNull();
+        expect(dropdown.nativeElement.classList).not.toContain('open');
     }
 }
 
 describe('components.bootstrap.dropdown', () => {
     describe('Dropdown', () => {
-        beforeEach(async(() => {
-            TestBed.configureTestingModule({
+        let fixture: ComponentFixture<DropdownHost>;
+        let dropdown: DebugElement;
+        let toggle: DebugElement;
+
+        beforeEach(async (done) => {
+            await TestBed.configureTestingModule({
                 declarations: [
                     CsDropdown,
                     CsDropdownToggle,
                     DropdownHost
                 ]
             }).compileComponents();
-        }));
 
-        let fixture: ComponentFixture<DropdownHost>;
-        let dropdown: DebugElement;
-        let toggle: DebugElement;
-
-        beforeEach(() => {
             fixture = TestBed.createComponent(DropdownHost);
             dropdown = fixture.debugElement.query(By.css('#hosted-dropdown'));
             toggle = fixture.debugElement.query(By.css('#dropdown-toggle'));
+            done();
         });
 
-        it('should open and close when _using component instance methods', async(() => {
+        it('should open and close when _using component instance methods', () => {
             fixture.detectChanges();
             dropdown.componentInstance.open();
             fixture.detectChanges();
@@ -75,9 +74,9 @@ describe('components.bootstrap.dropdown', () => {
             dropdown.componentInstance.close();
             fixture.detectChanges();
             expectIsDropdownOpen(dropdown, false);
-        }));
+        });
 
-        it('should toggle the dropdown if the toggle is clicked',async(() => {
+        it('should toggle the dropdown if the toggle is clicked',() => {
             fixture.detectChanges();
             toggle.triggerEventHandler('click', {});
             fixture.detectChanges();
@@ -87,8 +86,7 @@ describe('components.bootstrap.dropdown', () => {
             fixture.detectChanges();
 
             expectIsDropdownOpen(dropdown, false);
-
-        }));
+        });
 
     });
 });